Filling in missing values in dataset 2
Filled in a lot of missing values in database 2. It is one of the largest dataframes in our project and I managed to proces a lot of columns till now. I am now considering how to fill in the missing values for the city, state, zipcode and county data as they are tied between LA, CA and PHX, AZ.